+++ src/lib.rs
@@ -1,10 +1,12 @@
 mod api_client;
 mod cli;
+mod cli_types;
 mod error;
 mod serde;
 mod types;
 
 pub use api_client::*;
 pub use cli::*;
+pub use cli_types::*;
 pub use error::ApiError;
 pub use types::*;
